+++ src/pages/CartList/styles.js
@@ -2,8 +2,6 @@ import styled from '@emotion/styled/macro';
 
 import FlexContainer from 'components/@common/FlexContainer';
 
-import { COLORS } from 'styles/theme';
-
 const Container = styled.div`
   display: grid;
   grid-template-columns: 58% 40%;
@@ -14,18 +12,4 @@ const Container = styled.div`
   }
 `;
 
-const OrderContainer = styled.div`
-  display: flex;
-  position: sticky;
-  top: 3.125rem;
-
-  flex-direction: column;
-  border: 1px solid ${COLORS.GRAY_150};
-
-  ${FlexContainer} {
-    padding: 1.5rem;
-    font-weight: bold;
-  }
-`;
-
-export { Container, OrderContainer };
+export { Container };
